+++ scripts/variables.sh
@@ -1,0 +1,11 @@
+pipe_pane_key_option="@pipe-pane-key"
+default_pipe_pane_key="P" # Shift-p
+
+pane_screenshot_key_option="@pane-screenshot-key"
+default_pane_screenshot_key="M-p" # Alt-p
+
+scrollback_dump_key_option="@scrollback-dump-key"
+default_scrollback_dump_key="M-P" # Alt-Shift-p
+
+clear_scrollback_key_option="@scrollback-clear-key"
+default_clear_scrollback_key="M-c" # Alt-c